Adding PGPLOT to the SCHED distribution.  It has been the biggest source of frustration for users installing the code.  I plan to provide all that is needed for an installation, but not fully integrate it with SCHED in case the site has another installation.  Because of the environment variable PGPLOT_DIR I cannot make a totally separate installation that does not affect other users (unless I change that variable wherever it appears).   More integration will be needed to make sure all this works.
This is version 5.3.0 of pgplot.  If you go to the pgplot home page, you
will find that the latest release is 5.2.2.  I got 5.3.0 from Tim
Pearson many years ago when trying to install on Macs, for which it
has additional support.  It was not a public release, and it seems it 
never did get fully released.  However, I think the version on some of
the repositories is this version.


git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@340 d0039845-cc32-0410-9ae4-e027caa9de09