This includes additino of comments to the if_def lines in VEX to help the new VLBA control system set switches.
The beginnings of the ability to protect scans from preemption by
the USNO daily observations is there.

Some alterations of the DELZN segment generation is included.  Most
recent work in this area supports small numbers of stations and long
segments as needed by the USNO.

The Linux64gfortran compile option was improved.

There are catalog updates.

This is definitely an intermediate revision with significant 
partially complete modifications.  But it does run and pass 
the verify test.  A new Stdout was made using the results.