Oops, last checkin was premature (quit wrong editor window!). The note should have read:
Major revision to include a more or less full description of the DDC personality of the DBBC. The PFB personality is assumed to be the same as the RDBE, until I get some decent documentation. Some minor changes to LBA also. This all needs more testing, but appears to work.

- Added 'DBBCVER' keyword to station catalogue to cover the available DBBC versi
ons. Currently accepts 'astro', 'geo' or hybrid.

- Added bbclba.f and bbcdbbc.f routines to set default BBC->IF mappings for LBA and DBBC respectively. For dbbc the allowed mapping is set in ifdbbc.f so that it can also be called from chkdbbc.f.

- Modified settr.f and vexwrtr.f to respectively set and use the correct channel->bitstream (track) mappings for the MK5B and LBA recorders.

- fixed bug in writing data transfer commands (potential segmentation fault for first scan in schedule).




git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@468 d0039845-cc32-0410-9ae4-e027caa9de09