I am trying to make it possible to use masers for reference pointing, with Doppler frequency settings and narrow bandwidths in the legacy VLBA BBCs in observations that use the RDBE, especially with the PFB personality which has no tuning or bandwidth flexibility.  This triggered a need to deal with the pulse cal detection in the legacy system a the detector setup would track the legacy BBC settings.  That in turn triggered a restructuring of the handling of pulse cal setup information.  The pcal sets are now gone.  The pulse cal settings have been merged into the frequency sets and the legacy detector settings take a less prominent role - like they don't appear in the .sum file.
The new code passes Verify, but has not been tested for the new BBC
setting flexibility.  There were a lot of changes, including some in the
Vex code, so I did not want to sit on it until the new features are 
tested.

The catalog changes are mainly because of running Update with new
information from elsewhere.

Craig Walker  Sept. 17, 2013.



git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@526 d0039845-cc32-0410-9ae4-e027caa9de09