Pre shutdown commit to help with transfers to other machines.
This is a working version that includes support for reference pointing with
the VLBA legacy system when using the RDBE, especially the PFB version that 
can't tune, including the use of narrow band observations with masers using
Doppler calculations to get the frequencies.  While the timing of the commit
was set by Washington politics, this should generally be a useful version.

Craig  Oct 4, 2013



git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@527 d0039845-cc32-0410-9ae4-e027caa9de09