This commit includes:
The new version of the manual set up for htlatex and pdflatex.

Updated catalogs.

The ability to use TLE ephemeris files for satellite tracking.

Updated version of the locations and stations maintenance programs.

Various small changes throughout the code going back to mid November.

I haven't yet changed the version, but should do that soon.

I hope we can do a release fairly soon, but there will be a bit
more before we are ready for that.

Jan. 22, 2015  RC Walker.



git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@628 d0039845-cc32-0410-9ae4-e027caa9de09