This commit is mostly items related to catalogs.  I how have source catalogs being delivered from GSFC and Petrov that do not need modification.  I also have a locations catalog from GSFC that does not need modification.  I plan to abandon sources.vlba and am in the process of modifying the examples to use either sources.gsfc or sources.rfc (both of which are symlinks to the latest release from those groups (rfc is Radio Fundamental Catalog from Petrov). The new catalogs do not have quite the same stations as the old so many of the primary catslogs had stations added or subtracted to make a coherent whole.
There were minor changes to the code for plotting station positions for 
configuration searches.  There are partial updates to the manual and the 
release instructions.  

I'm working toward a release, but this commit was driven partly by a desire 
to move the work to my home machine.  It is not complete.

March 2, 2015  RC Walker


git-svn-id: https://svn.aoc.nrao.edu/repos/sched/trunk/sched@649 d0039845-cc32-0410-9ae4-e027caa9de09