Added satellite support when outputing .vex files.
Currently a .key file can specify a satellite source via something like:

satinit /
satname = 'prn08'  kerfile = 'current.tls' satnum = 40730 tlefile = 'prn008.tle' /
endsat /

However, this is ignored by sched when outputing to a .vex file.  The
new logic causes the vex statement:

source_type = prn008.tle : 40730;

to be output to the source definition of the .vex file.  This is
compatible with the parsing of the existing vex statement but
redefines the meaning; however, the VLBA does not appear to be using
the source_type statement in vex files.  With this addition to sched,
vex2script can use the information in the produced .vex files to
create scripts which will contain pointing information produced using
the orbit parameter file specified in the .key file.