SCHED12.0 introduces a VEX2 writer and control KEYword.
Partially copy this.

SCHED12.0 functionality:
VEXVRSN controls which VEX version(s) to write, default being 2.0.
If one VEX file is written, it is called <exp>.vex. For two files
.vex1 and .vex2 are used.

pySCHED:
VEXVRSN is introduced, but the default is 'both' for backward
compatibility. The VEX2 file will always be names <exp>.vex2.